vochain: move some methods from app.go into new cometbft.go
* Info
* InitChain
* CheckTx
* FinalizeBlock
* Commit
* PrepareProposal
* ProcessProposal
* ListSnapshots
* OfferSnapshot
* LoadSnapshotChunk
* ApplySnapshotChunk
* Query
* ExtendVote
* VerifyVoteExtension
